feat(auth): adds pages for login and signup
This commit adds pages with forms for loggin in and signing
up of users. These are added as individual components which
share a single module, the auth module.

I also added a service which holds the functionality for
logging in and signing up users to the backend of the application