feat(items): ability to retrieve and display items
This commit adds the ability to retrieve items from the
backend and display them in the Items Component.

It also gives the ability to filter items based on the
items reporter, either finder or owner to give results
for items that have been reported lost or those that have
been reported found and not picked up.