feat(authentication): adds auth emitters
This commit adds authentication emitters that emit an
event on login and logout. It also adds listeners within
the navbar to listen to determine whetehr to display
the login button or the user name.

It changes the folder title for services from _services
to services.

It adds an extra button class for primary light classes
effectively changing the color of the authentication action
buttons to the $app-blue color.

It provides notification ability to alert the user if their
logout action is successful.