feat(add-item): adds modal to add item.
This commit adds a modal that will be used to add items.
It provides a user with the ability to set the name, category
description of an item within the front end. No item is actually
added when user adds an item, just the modal view has been built.

It also adds a couple of reusable components for the form select,
input and textarea elements.